backend/app/requirements.txt
cassandra-driver==3.29.0
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.29.0
pydantic
python-dotenv==1.0.1
httpx==0.27.0
openai==0.28.1
PyJWT==2.8.0
prometheus-client==0.20.0
mypy==1.10.0
slowapi==0.1.8
redis==5.0.3